+++ app/controllers/users_controller.rb
@@ -17,6 +17,10 @@ class UsersController < ApplicationController
     @comments = @user.comments
   end
 
+  def index
+    @users = User.top_rated.paginate(page: params[:page], per_page: 10)
+  end
+
   private
 
   def user_params
