+++ app/views/layouts/application.html.erb
@@ -12,6 +12,9 @@
     <ul class="nav nav-tabs">
       <li><%= link_to "Home", root_path %></li>
       <li><%= link_to "About", about_path %></li>
+      <li><%= link_to "Topics", topics_path %></li>
+      <li><%= link_to "Popular", posts_path %></li>
+      <li><%= link_to "Users", users_path %></li>
 
       <div class="pull-right user-info">
         <% if current_user %>
