+++ app/views/topics/posts/show.html.erb
