+++ app/views/users/_user.html.erb
@@ -1,0 +1,14 @@
+<div class="media">
+  <%= link_to user, class: 'media-left' do %>
+    <%= image_tag(user.avatar.profile.url, class: "media-object") if user.avatar? %>
+  <% end %>
+  <div class="media-body">
+    <h4 class="media-heading">
+      <%= link_to user.name, user %>
+    </h4>
+    <small>
+      <%= pluralize(user.posts_count, 'post') %>,
+      <%= pluralize(user.comments_count, 'comment') %>
+    </small>
+  </div>
+</div>
